requirements.txt
pandas 
numpy
scikit-learn==1.5.2
matplotlib
seaborn
sqlalchemy
mysql-connector-python
python-dotenv
pymysql
